fix ptr to int warning on 32bit platforms

+++ C/include/obl_device.h
@@ -326,9 +326,9 @@ static int obl_main(){
 	    case OBL_CMD_INFO:
 	      obl_send_status(ok);error=OBL_SUCCESS;
 	      obl_send_str(" stack_buf=0x");
-	      obl_send_hex64((uint64_t)stack_buf);
+	      obl_send_hex64((uint64_t)(uintptr_t)stack_buf);
 	      obl_send_str(", obl_main=0x");
-	      obl_send_hex64((uint64_t)obl_main);
+	      obl_send_hex64((uint64_t)(uintptr_t)obl_main);
 	      break;
 	    }
 		}while(0);
